fixup! fixup! fixup! fixup! fixup! fixup! fixup! fixup! TW-2796: Implement autoplay for voice messages and enhance audio event handling
